requirements.txt
Django>=4.0
djangorestframework
channels>=4.0
channels-redis>=4.0
